+++ epserde/src/impls/string.rs
@@ -51,9 +51,9 @@ impl SerInner for String {
 
 impl DeserInner for String {
     fn __check_covariance<'__long: '__short, '__short>(
-        p: deser::CovariantProof<Self::DeserType<'__long>>,
+        proof: deser::CovariantProof<Self::DeserType<'__long>>,
     ) -> deser::CovariantProof<Self::DeserType<'__short>> {
-        p
+        proof
     }
     unsafe fn _deser_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<Self> {
         let slice = unsafe { deser_full_vec_zero(backend) }?;
@@ -102,9 +102,9 @@ impl SerInner for Box<str> {
 
 impl DeserInner for Box<str> {
     fn __check_covariance<'__long: '__short, '__short>(
-        p: deser::CovariantProof<Self::DeserType<'__long>>,
+        proof: deser::CovariantProof<Self::DeserType<'__long>>,
     ) -> deser::CovariantProof<Self::DeserType<'__short>> {
-        p
+        proof
     }
     #[inline(always)]
     unsafe fn _deser_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<Self> {