+++ epserde/src/impls/tuple.rs
@@ -86,9 +86,9 @@ macro_rules! impl_tuples {
 
 		impl<T: ZeroCopy> DeserInner for ($($t,)*) {
             fn __check_covariance<'__long: '__short, '__short>(
-                p: deser::CovariantProof<Self::DeserType<'__long>>,
+                proof: deser::CovariantProof<Self::DeserType<'__long>>,
             ) -> deser::CovariantProof<Self::DeserType<'__short>> {
-                p
+                proof
             }
             type DeserType<'a> = &'a ($($t,)*);
             unsafe fn _deser_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<Self> {